+++ test/core/test_git.py
@@ -24,8 +24,7 @@ from iambic.core.git import (
     get_origin_head,
     get_remote_default_branch,
 )
-from iambic.core.models import BaseTemplate
-from iambic.core.template import ConfigMixin
+from iambic.core.models import BaseTemplate, ConfigMixin
 from iambic.plugins.v0_1_0.example.local_file.models import (
     ExampleLocalFileMultiAccountTemplate,
     ExampleLocalFileMultiAccountTemplateProperties,
@@ -417,7 +416,7 @@ def test_create_templates_for_deleted_files(mocker):
 
     deleted_files = [git_diff1, git_diff2]
     mocker.patch(
-        "iambic.core.template.ConfigMixin.templates",
+        "iambic.core.models.ConfigMixin.templates",
         new_callable=mocker.PropertyMock,
         return_value=[MockTemplate, MockTemplate2],
     )
